requirements.txt
FunPayAPI==1.1.0
python-telegram-bot==22.5
steampy==1.2.0
steam==1.4.4
PyYAML==6.0.3
protobuf==6.33.1
fastapi==0.121.3
uvicorn==0.38.0